refactor: reorganize ownable-std module layout
